apps/api/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "pydantic-settings>=2.10.1",
    "email-validator>=2.2.0",
    # Database
    "sqlalchemy>=2.0.43",
    "psycopg[binary]>=3.2.3",
    "alembic>=1.16.5",
    "greenlet>=3.1.1",
    # Authentication
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.19",
    # Data processing (for MMM model) - versions compatible with google-meridian
    "pandas>=2.2.3",
    "numpy>=1.26.0,<2.1.0",
    "scikit-learn>=1.5.0",
    "scipy>=1.12.0,<1.13",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    "xarray>=2024.10.0",
    "joblib>=1.4.0",
    # Google Meridian MMM package
    "google-meridian>=1.0.0",
    # Development
    "python-dotenv>=1.0.1",
    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "pytest-mock>=3.12.0",
    "coverage>=7.4.0",
    "aiosqlite>=0.20.0",
]
